[ENGOPS-1125] Refactor emr32 to emr34 #3 more package paths.
